Create README.md
This is a full stack application designed to be a fully functional website for commercial use. Legacy Builds LLC. is my active custom computer business where we build, repair, and service computers and related peripherals. This is going to be deployed via AWS. Created with MERN stack. CRUD user functionality.